+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CoreExtensionsKtTest.kt
@@ -26,7 +26,6 @@ class CoreExtensionsKtTest {
 	fun nearestStackInfo() {
 		DELAY()
 		FIXME()
-		TODO()
 	}
 
 	@Test